refactor: 구간 관리 controller , service, test  분리
