+++ src/test/java/wooteco/subway/acceptance/SectionAcceptanceTest.java
@@ -1,0 +1,119 @@
+package wooteco.subway.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import wooteco.subway.dto.LineRequest;
+import wooteco.subway.dto.SectionRequest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class SectionAcceptanceTest extends AcceptanceTest {
+
+    @Test
+    @DisplayName("새로운 구간을 추가한다.(맨앞)")
+    void addSection_First() {
+        // given
+        createLineResponse(new LineRequest("2호선", "green", 1L, 2L, 10));
+        SectionRequest sectionRequest = new SectionRequest(3L, 1L, 5);
+
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .body(sectionRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/lines/1/sections")
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    @Test
+    @DisplayName("새로운 구간을 추가한다.(사이)")
+    void addSection_Middle() {
+        // given
+        createLineResponse(new LineRequest("2호선", "green", 1L, 2L, 10));
+        SectionRequest sectionRequest = new SectionRequest(1L, 3L, 5);
+
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .body(sectionRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/lines/1/sections")
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    @Test
+    @DisplayName("새로운 구간을 추가한다.(맨뒤)")
+    void addSection_End() {
+        // given
+        createLineResponse(new LineRequest("2호선", "green", 1L, 2L, 10));
+        SectionRequest sectionRequest = new SectionRequest(2L, 3L, 5);
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .body(sectionRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/lines/1/sections")
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    @Test
+    @DisplayName("구간을 삭제한다.")
+    void deleteSection() {
+        // given
+        createLineResponse(new LineRequest("2호선", "green", 1L, 2L, 10));
+        SectionRequest sectionRequest = new SectionRequest(2L, 3L, 5);
+        RestAssured.given().log().all()
+                .body(sectionRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .post("/lines/1/sections")
+                .then().log().all()
+                .extract();
+
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .delete("/lines/1/sections?stationId=2")
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    @Test
+    @DisplayName("구간이 하나일경우 예외를 발생시킨다.")
+    void deleteSection_Exception() {
+        // given
+        createLineResponse(new LineRequest("2호선", "green", 1L, 2L, 10));
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .delete("/lines/1/sections?stationId=2")
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+
+}
